# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: AmexingWeb API
  description: |
    PCI DSS compliant e-commerce platform API built with Parse Server and Express.js.
    
    ## Authentication
    This API uses session-based authentication with Parse Server. Include session tokens in the `X-Parse-Session-Token` header.
    
    ## Rate Limiting
    API endpoints are rate-limited to prevent abuse. Check response headers for rate limit information.
    
    ## Security
    All endpoints enforce HTTPS in production and include comprehensive security headers.
  version: 1.0.0
  contact:
    name: AmexingWeb Team
    email: support@amexingweb.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:1337
    description: Development server
  - url: https://api.amexingweb.com
    description: Production server

tags:
  - name: Authentication
    description: User authentication and session management
  - name: Health
    description: System health and monitoring endpoints
  - name: Users
    description: User management operations
  - name: Products
    description: Product catalog operations

paths:
  /health:
    get:
      tags:
        - Health
      summary: System health check
      description: Returns the current health status of the application
      operationId: getHealth
      responses:
        '200':
          description: System is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
        '503':
          description: System is unhealthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /metrics:
    get:
      tags:
        - Health
      summary: System metrics
      description: Returns system performance metrics and statistics
      operationId: getMetrics
      responses:
        '200':
          description: System metrics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MetricsResponse'

  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user with username/email and password
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too many login attempts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/auth/register:
    post:
      tags:
        - Authentication
      summary: User registration
      description: Register a new user account
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Invalid request data or user already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/auth/logout:
    post:
      tags:
        - Authentication
      summary: User logout
      description: Logout current user and invalidate session
      operationId: logout
      security:
        - sessionToken: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/users/me:
    get:
      tags:
        - Users
      summary: Get current user
      description: Retrieve information about the currently authenticated user
      operationId: getCurrentUser
      security:
        - sessionToken: []
      responses:
        '200':
          description: User information retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    sessionToken:
      type: apiKey
      in: header
      name: X-Parse-Session-Token
      description: Parse Server session token

  schemas:
    HealthResponse:
      type: object
      required:
        - status
        - timestamp
        - uptime
        - environment
      properties:
        status:
          type: string
          enum: [healthy, unhealthy]
          example: healthy
        timestamp:
          type: string
          format: date-time
          example: "2025-08-19T15:30:00.000Z"
        uptime:
          type: number
          description: Server uptime in seconds
          example: 3600.5
        environment:
          type: string
          example: development
        database:
          type: object
          properties:
            connected:
              type: boolean
              example: true
            responseTime:
              type: number
              description: Database response time in milliseconds
              example: 25

    MetricsResponse:
      type: object
      properties:
        system:
          type: object
          properties:
            uptime:
              type: number
              example: 3600.5
            memory:
              type: object
              properties:
                used:
                  type: number
                  example: 52428800
                total:
                  type: number
                  example: 134217728
            cpu:
              type: object
              properties:
                usage:
                  type: number
                  example: 15.5
        application:
          type: object
          properties:
            activeConnections:
              type: number
              example: 12
            requestsPerMinute:
              type: number
              example: 450
            averageResponseTime:
              type: number
              example: 125

    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          description: Username or email address
          example: john.doe@example.com
        password:
          type: string
          format: password
          minLength: 8
          example: SecurePass123!

    RegisterRequest:
      type: object
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 20
          pattern: '^[a-zA-Z0-9_]+$'
          example: johndoe
        email:
          type: string
          format: email
          example: john.doe@example.com
        password:
          type: string
          format: password
          minLength: 12
          description: Must contain uppercase, lowercase, numbers, and special characters
          example: SecurePass123!
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe

    LoginResponse:
      type: object
      required:
        - success
        - user
        - sessionToken
      properties:
        success:
          type: boolean
          example: true
        user:
          $ref: '#/components/schemas/User'
        sessionToken:
          type: string
          example: r:1234567890abcdef

    UserResponse:
      type: object
      required:
        - success
        - user
      properties:
        success:
          type: boolean
          example: true
        user:
          $ref: '#/components/schemas/User'

    User:
      type: object
      required:
        - objectId
        - username
        - email
        - createdAt
        - updatedAt
      properties:
        objectId:
          type: string
          example: abc123def456
        username:
          type: string
          example: johndoe
        email:
          type: string
          format: email
          example: john.doe@example.com
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        emailVerified:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
          example: "2025-08-19T15:30:00.000Z"
        updatedAt:
          type: string
          format: date-time
          example: "2025-08-19T15:30:00.000Z"

    SuccessResponse:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation completed successfully

    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: Invalid credentials
        message:
          type: string
          example: The provided username or password is incorrect
        code:
          type: integer
          example: 401
        details:
          type: object
          description: Additional error details

security:
  - sessionToken: []